feat: 완주하지 못한 선수
