+++ src/main/java/junyoung/week21/Week21JeongjySolution.java
@@ -1,5 +1,8 @@
 package junyoung.week21;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class Week21JeongjySolution {
 
     public int splitString(String s) {
@@ -28,5 +31,22 @@ public class Week21JeongjySolution {
             answer++;
         }
         return answer;
+    }
+
+    public String nonFinisher(String[] participant, String[] completion) {
+        Map<String, Integer> nameToCount = new HashMap<>();
+        for (String p : participant) {
+            nameToCount.put(p, nameToCount.getOrDefault(p, 0) + 1);
+        }
+
+        for (String c : completion) {
+            nameToCount.compute(c, (k, v) -> v - 1);
+        }
+        return nameToCount.entrySet()
+                .stream()
+                .filter(s -> s.getValue() > 0)
+                .findFirst()
+                .get()
+                .getKey();
     }
 }
